Added Profile picture loading
Added a few images to test if the profile picture stuff works.
There is only one issue, there is no way to upload profile pictures yet

How it works:
There are two JS files that can be called dependent on what is needed. It takes a user id as an input and gives an image that as a name has this id. First it tries a PNG then a JPG and after that it takes a default-avatar. Profile pictures are directly saved in a dedicated public folder.